RAGevaluator/requirements.txt
pandas
streamlit
chardet
python-dotenv
langchain
langchain-community
langchain-core
langchain-openai
openai
